Refactor review functions into new object
